updateจ้า

+++ src/users/users.service.ts
@@ -10,24 +10,42 @@ export class UsersService {
   constructor(
     @InjectRepository(User)
     private usersRepository: Repository<User>,
-  ) {}
-  create(createUserDto: CreateUserDto) {
-    return this.usersRepository.save(createUserDto);
+  ) { }
+  async create(createUserDto: CreateUserDto) {
+    const user = await this.usersRepository.create(createUserDto);
+    await this.usersRepository.save(user);
+    return user;
   }
 
   findAll(): Promise<User[]> {
-    return this.usersRepository.find();
+    return this.usersRepository.find
+      (
+        {
+          select:
+            [
+              'firstName', 'lastName',
+            ],
+          order:{
+            id: 'DESC'
+          }
+        }
+      );
   }
 
   findOne(id: number): Promise<User | null> {
     return this.usersRepository.findOneBy({ id });
   }
 
-  update(id: number, updateUserDto: UpdateUserDto) {
-    return this.usersRepository.update(id, updateUserDto);
+  async update(id: number, updateUserDto: UpdateUserDto) {
+    await this.usersRepository.update(id, updateUserDto);
+    return this.usersRepository.findOneBy({ id });
   }
 
-  async remove(id: number): Promise<void> {
-    await this.usersRepository.delete(id);
+  async remove(id: number) {
+    const user = await this.usersRepository.findOneBy({ id });
+    if (user == null) {
+      return null;
+    }
+    return await this.usersRepository.delete(id);
   }
 }
